Pass callbackContext to each function
For most functions I have simply passed them the callbackContext to use
instead of the class instance variable. I’ve renamed the class instance
variable instanceCallbackContext to make this clearer.

In createCalendar(), deleteEvent() and findEvents(), I have removed the
call to PluginResult.setKeepCalback(true). This is what was filling the
cordova.callbacks hash. I cannot see why this was being used. What have
I missed?

I have not made this change for functions that use the
startActivityForResult() —> onActivityResult() flow. Not sure what can
be done for these. Can the callbackContext be passed to
onActivityResult() via the Intent class? Also, as these functions also
call the callbackContext, should they not be using
PluginResult.setKeepCallback(true) if onActivityResult() is going to
call it as well?